refactoring without singleton pattern
