+++ webauthn4j-core/src/main/java/com/webauthn4j/converter/util/CborConverter.java
@@ -17,31 +17,16 @@
 package com.webauthn4j.converter.util;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import com.fasterxml.jackson.dataformat.cbor.CBORFactory;
-import com.webauthn4j.converter.exception.DataConversionException;
 import com.webauthn4j.converter.jackson.WebAuthnCBORModule;
-import com.webauthn4j.data.extension.authenticator.AuthenticationExtensionsAuthenticatorOutputs;
 import com.webauthn4j.util.AssertUtil;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UncheckedIOException;
 
 /**
  * A utility class for CBOR serialization/deserialization
  */
-public class CborConverter {
-    private static final String INPUT_MISMATCH_ERROR_MESSAGE = "Input data does not match expected form";
-
-    private ObjectMapper cborMapper;
-
+public class CborConverter extends AbstractCborConverter {
     public CborConverter(ObjectMapper cborMapper) {
         AssertUtil.notNull(cborMapper, "cborMapper must not be null");
 
@@ -59,72 +44,4 @@ public class CborConverter {
         this(new ObjectMapper(new CBORFactory()));
     }
 
-    @SuppressWarnings("unchecked")
-    public <T> T readValue(byte[] src, Class valueType) {
-        try {
-            return (T) getCborMapper().readValue(src, valueType);
-        } catch (MismatchedInputException | JsonParseException e) {
-            throw new DataConversionException(INPUT_MISMATCH_ERROR_MESSAGE, e);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> T readValue(InputStream src, Class valueType) {
-        try {
-            return (T) getCborMapper().readValue(src, valueType);
-        } catch (MismatchedInputException | JsonParseException e) {
-            throw new DataConversionException(INPUT_MISMATCH_ERROR_MESSAGE, e);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public <T> T readValue(byte[] src, TypeReference valueTypeRef) {
-        try {
-            return getCborMapper().readValue(src, valueTypeRef);
-        } catch (MismatchedInputException | JsonParseException e) {
-            throw new DataConversionException(INPUT_MISMATCH_ERROR_MESSAGE, e);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public AuthenticationExtensionsAuthenticatorOutputs readValue(InputStream inputStream, TypeReference<AuthenticationExtensionsAuthenticatorOutputs> typeReference) {
-        try {
-            return getCborMapper().readValue(inputStream, typeReference);
-        } catch (MismatchedInputException | JsonParseException e) {
-            throw new DataConversionException(INPUT_MISMATCH_ERROR_MESSAGE, e);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public JsonNode readTree(byte[] bytes) {
-        try {
-            return getCborMapper().readTree(bytes);
-        } catch (MismatchedInputException | JsonParseException e) {
-            throw new DataConversionException(INPUT_MISMATCH_ERROR_MESSAGE, e);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    public byte[] writeValueAsBytes(Object value) {
-        try {
-            return getCborMapper().writeValueAsBytes(value);
-        } catch (JsonProcessingException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    /**
-     * Returns the {@link ObjectMapper} configured for CBOR processing
-     *
-     * @return the {@link ObjectMapper} configured for CBOR processing
-     */
-    private ObjectMapper getCborMapper() {
-        return cborMapper;
-    }
 }
