+++ webauthn4j-core/src/main/java/com/webauthn4j/converter/util/JsonConverter.java
@@ -16,30 +16,19 @@
 
 package com.webauthn4j.converter.util;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.cbor.CBORFactory;
 import com.webauthn4j.converter.jackson.WebAuthnJSONModule;
-import com.webauthn4j.util.AssertUtil;
 
 /**
  * A utility class for JSON serialization/deserialization
  */
 public class JsonConverter extends AbstractJsonConverter {
     public JsonConverter(ObjectMapper jsonMapper) {
-        AssertUtil.notNull(jsonMapper, "jsonMapper must not be null");
-
-        AssertUtil.isTrue(!(jsonMapper.getFactory() instanceof CBORFactory), "factory of jsonMapper must be JsonFactory.");
-
-        this.jsonMapper = jsonMapper;
-
-        this.jsonMapper.registerModule(new WebAuthnJSONModule());
-        this.jsonMapper.configure(DeserializationFeature.WRAP_EXCEPTIONS, false);
-        this.jsonMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        this.jsonMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-
+        super(jsonMapper,
+                new WebAuthnJSONModule(),
+                false,
+                false);
     }
 
     public JsonConverter() {
